trying to come up with an editor ...
I briefly looked at raygui for immediate UI usage, but there are several things I am not content with; overlapping UI does not block clicks, the styles are limited to filled rectangles... otherwise it would be neat. So ... I am trying to come up with my own imgui-ui system that tries to handle this the way I want it to work.